[refacto] Rename preview app
