+++ django/river/urls.py
@@ -1,10 +1,9 @@
-from django.contrib import admin
 from django.urls import include, path
 
 from core import urls as core_urls
-from preview import urls as preview_urls
+from control import urls as control_urls
 
 urlpatterns = [
     path("api/core/", include(core_urls)),
-    path("api/", include(preview_urls)),
+    path("api/", include(control_urls)),
 ]
